+++ R/clustering.R
@@ -3309,20 +3309,31 @@ getDendrogramSplits <- function(
 #' one of the sets to the other based on KNN similarity voting in that space.
 #' @param x target object
 #' @param y source object
+#' @param spat_unit spatial unit. A character vector of 2 can also be passed
+#' for x (1) and y (2). Setting defaults with `activeSpatUnit()` may be easier
+#' @param feat_type feature type. A character vector of 2 can also be passed
+#' for x (1) and y (2). Setting defaults with `activeFeatType()` may be easier
 #' @param source_cell_ids cell/spatial IDs with the source labels to transfer
 #' @param target_cell_ids cell/spatial IDs to transfer the labels to.
 #' IDs from `source_cell_ids` are always included as well.
 #' @param labels metadata column in source with labels to transfer
 #' @param k number of k-neighbors to train a KNN classifier
 #' @param name metadata column in target to apply the full set of labels to
+#' @param integration_method character. Integration method to use when
+#' transferring labels. Options are "none" (default) and "harmony". See section
+#' below for more info and params.
 #' @param prob output knn probabilities together with label predictions
 #' @param reduction reduction on cells or features (default = "cells")
 #' @param reduction_method shared reduction method (default = "pca" space)
 #' @param reduction_name name of shared reduction space (default name = "pca")
 #' @param dimensions_to_use dimensions to use in shared reduction space
 #' (default = 1:10)
-#' @returns object `x` with new transferred labels added to metadata
 #' @inheritDotParams FNN::knn -train -test -cl -k -prob
+#' @returns object `x` with new transferred labels added to metadata. If
+#' running on `x` and `y` objects, `integration_method = "harmony"`,
+#' `plot_join_labels = TRUE`, and `return_plot = TRUE` is set, output will
+#' be instead a named list of `gobject` (updated `x`), and `label_source_plot`
+#' and `label_target_plot` `ggplot2` objects
 #' @details
 #' This function trains a KNN classifier with [FNN::knn()].
 #' The training data is from object `y` or `source_cell_ids` subset in `x` and
@@ -3342,6 +3353,34 @@ getDendrogramSplits <- function(
 #' used to transfer labels from one set of annotated data to another dataset
 #' based on expression similarity after joining and integrating.
 #'
+#' # integration_method
+#' When running `labelTranfer()` on two `giotto` objects, an integration
+#' pipeline can also be run to align the two datasets together before the
+#' transfer. `integration_method = "harmony"` will make a temporary joined 
+#' object on shared features, filter to remove 0 values, run PCA, then harmony
+#' integration, before performing the label transfer from `y` to `x` on the
+#' integrated harmony embedding space. Additional params that can be used with
+#' this method are:\cr
+#' 
+#' * `source_cell_ids` - character. subset of `y` cells to use
+#' * `target_cell_ids` - character. subset of `x` cells to use
+#' * `expression_values` - character. expression values in `x` and `y` to use 
+#' to generate combined space. Default = `"raw"`
+#' * `use_hvf` - logical. whether to calculate highly variable features to use 
+#' for PCA calculation. Default = `TRUE`, but setting `FALSE` is recommended if
+#' any of `x` or `y` has roughly 1000 features or fewer
+#' * `plot_join_labels` - logical. Whether to plot source labels and final
+#' labels in the joine object UMAP.
+#' * `normalize_params` - named list. Additional params to pass to 
+#' `normalizeGiotto()` if desired.
+#' * `pca_params` - named list. Additional params to pass to `runPCA()` if
+#' desired.
+#' * `integration_params` - named list. Additional params to pass to 
+#' `runGiottoHarmony()` if desired.
+#' * `plot_params` - named list. Additional params to pass to `plotUMAP()` if
+#' desired. Only relevant when `plot_join_labels = TRUE`
+#' * `verbose` - verbosity
+#' 
 #' @examples
 #' g <- GiottoData::loadGiottoMini("visium")
 #' id_subset <- sample(spatIDs(g), 300)
@@ -3364,45 +3403,314 @@ setGeneric(
     function(x, y, ...) standardGeneric("labelTransfer")
 )
 
+
+.lab_transfer_harmony <- function(x, y,
+    source_cell_ids = NULL,
+    target_cell_ids = NULL,
+    expression_values = "raw",
+    dimensions_to_use = 1:10,
+    spat_unit = NULL,
+    feat_type = NULL,
+    use_hvf = TRUE,
+    plot_join_labels = FALSE,
+    normalize_params = list(),
+    pca_params = list(),
+    integration_params = list(),
+    plot_params = list(),
+    verbose = NULL,
+    ... # passes to labelTransfer
+) {
+    # NSE vars
+    cell_ID <- transfer <- transfer_prob <- NULL
+
+    .check_plot_output <- function(pparam) {
+        direct_setting <- plot_params[[pparam]]
+        if (!is.null(direct_setting)) return(direct_setting)
+        any(instructions(x, pparam), instructions(y, pparam))
+    }
+    
+    .show_plot <- .check_plot_output("show_plot")
+    .return_plot <- .check_plot_output("return_plot")
+    .save_plot <- .check_plot_output("save_plot")
+    
+    # turn off lower level updates
+    options("giotto.update_param" = FALSE)
+    on.exit({
+        options("giotto.update_param" = TRUE)
+    }, add = TRUE)
+        
+    vmsg(.v = verbose, "1. Creating temporary joined object...")
+    # match features
+    ufids <- intersect(featIDs(x), featIDs(y))
+    if (length(ufids) == 0L) {
+        stop("labelTransfer harmony: No common features between `x` and `y`",
+             call. = FALSE)
+    }
+
+    # get needed subobjects
+    xdata <- x[[
+        "expression", 
+        expression_values, 
+        spat_unit = spat_unit[[1]],
+        feat_type = feat_type[[1]],
+    ]]
+    ydata <- y[[
+        "expression", 
+        expression_values, 
+        spat_unit = spat_unit[[2]],
+        feat_type = feat_type[[2]],
+    ]]
+    ymeta <- y[[
+        "cell_metadata",
+        spat_unit = spat_unit[[2]],
+        feat_type = feat_type[[2]]
+    ]]
+    
+    # harmonize nesting
+    objName(xdata) <- rep("raw", length(xdata))
+    objName(ydata) <- rep("raw", length(ydata))
+    ydata <- c(ydata, ymeta)
+    spatUnit(xdata) <- rep("cell", length(xdata))
+    featType(xdata) <- rep("rna", length(xdata))
+    spatUnit(ydata) <- rep("cell", length(ydata))
+    featType(ydata) <- rep("rna", length(ydata))
+    
+    dummy_sl_x <- data.table::data.table(
+        cell_ID = colnames(xdata[[1]][]),
+        sdimx = 0, sdimy = 0
+    ) |>
+        createSpatLocsObj(
+            name = "raw", spat_unit = "cell", verbose = FALSE
+        )
+    dummy_sl_y <- data.table::data.table(
+        cell_ID = colnames(ydata[[1]][]),
+        sdimx = 0, sdimy = 0
+    ) |>
+        createSpatLocsObj(
+            name = "raw", spat_unit = "cell", verbose = FALSE
+        )
+    dummy_instrs <- instructions(x)
+    
+    xdata <- c(xdata, dummy_sl_x)
+    ydata <- c(ydata, dummy_sl_y)
+    
+    # generate temp objects
+    xj <- setGiotto(giotto(instructions = dummy_instrs, initialize = FALSE), 
+                    xdata, verbose = FALSE)
+    yj <- setGiotto(giotto(instructions = dummy_instrs, initialize = FALSE), 
+                    ydata, verbose = FALSE)
+    
+    xj <- subsetGiotto(xj, feat_ids = ufids, cell_ids = target_cell_ids)
+    yj <- subsetGiotto(yj, feat_ids = ufids, cell_ids = source_cell_ids)
+    
+    # join on intersected feats and specified cell_IDs
+    j <- joinGiottoObjects(list(xj, yj), gobject_names = c("x", "y"))
+    
+    # cleanup
+    rm(y, xj, yj, xdata, ydata, ymeta, dummy_sl_x, dummy_sl_y, dummy_instrs)
+
+    vmsg(.v = verbose, "2. Performing simple filter...")
+    j <- filterGiotto(j, 
+        expression_threshold = 1, 
+        min_det_feats_per_cell = 1, 
+        feat_det_in_min_cells = 1,
+        verbose = FALSE
+    )
+    
+    vmsg(.v = verbose, "3. Running normalize...")
+    normalize_params$verbose <- FALSE
+    # process
+    j <- do.call(normalizeGiotto, 
+                 args = c(list(gobject = j), normalize_params)
+    )
+    
+    if (use_hvf) {
+        vmsg(.v = verbose, "-- Calculating HVF...")
+        j <- calculateHVF(j)
+        pca_params$feats_to_use = pca_params$feats_to_use %null% "hvf"
+    } else {
+        pca_params <- c(pca_params, list(feats_to_use = NULL))
+    }
+    
+    vmsg(.v = verbose, "4. Running PCA...")
+    j <- do.call(runPCA, args = c(list(gobject = j), pca_params))
+    
+    # TODO determine dims to use via cumvar >= 30%
+    
+    # harmony
+    integration_params$name <- "harmony"
+    integration_params$vars_use = "list_ID"
+    integration_params$dim_reduction_name = "pca"
+    integration_params$dimensions_to_use = dimensions_to_use
+    
+    vmsg(.v = verbose, "5. Generating shared Harmony embedding space...")
+    j <- do.call(runGiottoHarmony,
+                 c(list(gobject = j), integration_params)
+    )
+    
+    # transfer
+    transfer_args <- list(...)
+    transfer_args$x <- j
+    transfer_args$source_cell_ids <- spatIDs(j, subset = list_ID == "y")
+    transfer_args$dimensions_to_use = dimensions_to_use
+    transfer_args$reduction = "cells"
+    transfer_args$reduction_method = "harmony"
+    transfer_args$reduction_name = "harmony"   
+    
+    vmsg(.v = verbose, "6. Performing label transfer...")
+    j <- do.call(labelTransfer, transfer_args)
+    
+    res <- pDataDT(j)
+    res <- res[list_ID == "x"]
+    res[, cell_ID := gsub("^x-", "", cell_ID)]
+    tname <- transfer_args$name
+    res <- res[, .SD, .SDcols = c("cell_ID", tname, paste0(tname, "_prob"))]
+
+    x <- addCellMetadata(x,
+        new_metadata = res, 
+        by_column = TRUE, 
+        column_cell_ID = "cell_ID"
+    )
+    
+    # plot outputs
+    if (!any(.save_plot, .return_plot, .show_plot) && plot_join_labels) {
+        return(x) # return early if plotting not needed
+    }
+    
+    vmsg(.v = verbose, "7. plot_join_labels = TRUE: Running UMAP...")
+    
+    j <- runUMAP(j, 
+        dim_reduction_to_use = "harmony",
+        dim_reduction_name = "harmony",
+        name = "harmony_umap",
+        dimensions_to_use = dimensions_to_use
+    )
+    
+    plotlabs <- unique(res[[tname]])
+    default_colors <- getRainbowColors(
+        n = length(plotlabs), slim = c(0.5, 1), vlim = c(0.3, 1)
+    )
+    names(default_colors) <- plotlabs
+    
+    plot_params$gobject <- j
+    plot_params$dim_reduction_name <- "harmony_umap"
+    plot_params$color_as_factor <- TRUE
+    plot_params$point_size <- plot_params$point_size %null% 0.5
+    plot_params$point_border_stroke <- plot_params$point_border_stroke %null% 0
+    plot_params$cell_color_code <-
+        plot_params$cell_color_code %null% default_colors
+        
+    p1_params <- p2_params <- plot_params
+    
+    p1_params$cell_color <- transfer_args$labels
+    p1_params$select_cells <- spatIDs(j, subset = list_ID == "y")
+    p1_params$other_point_size <- p1_params$other_point_size %null% 0.3
+    p1_params$show_plot <- FALSE
+    p1_params$return_plot <- TRUE
+    p1_params$title <- "source labels"
+    
+    p2_params$cell_color <- tname
+    p2_params$show_plot <- FALSE
+    p2_params$return_plot <- TRUE
+    p2_params$title <- "final labels"
+    
+    plot_list <- list(label_source_plot = do.call(plotUMAP, p1_params), 
+                      label_final_plot = do.call(plotUMAP, p2_params))
+    
+    if (.show_plot) {
+        print(plot_grid(plotlist = plot_list))
+    } 
+    if (.save_plot) {
+        plot_output_handler(x, plot_list[[1]],
+            save_plot = .save_plot,
+            default_save_name = "transferLabels_source",
+        )
+    }
+    if (.return_plot) {
+        x <- c(list(gobject = x), plot_list)
+    }
+    
+    return(x)
+}
+
+
+# ** giotto, giotto ####
+
 #' @rdname labelTransfer
 #' @export
 setMethod("labelTransfer", signature(x = "giotto", y = "giotto"), function(
         x, y,
-        spat_unit = NULL,
-        feat_type = NULL,
         labels,
         k = 10,
         name = paste0("trnsfr_", labels),
+        integration_method = c("none", "harmony"),
         prob = TRUE,
         reduction = "cells",
         reduction_method = "pca",
         reduction_name = "pca",
         dimensions_to_use = 1:10,
+        spat_unit = NULL,
+        feat_type = NULL,
         return_gobject = TRUE,
         ...) {
-    # NSE vars
-    temp_name <- cell_ID <- temp_name_prob <- NULL
 
     package_check(pkg_name = "FNN", repository = "CRAN")
-    spat_unit <- set_default_spat_unit(x, spat_unit = spat_unit)
-    feat_type <- set_default_feat_type(x,
-        spat_unit = spat_unit, feat_type = feat_type
+    
+    integration_method <- match.arg(integration_method, choices = c(
+        "none", "harmony"
+    ))
+    
+    if (!labels %in% colnames(pDataDT(y))) {
+        stop("`labels` not found in cell metadata of `y`", call. = FALSE)
+    }
+    
+    # norm su and ft lengths
+    if (length(spat_unit) == 1L) {
+        spat_unit <- rep(spat_unit, 2L)
+    }
+    if (length(feat_type) == 1L) {
+        feat_type <- rep(feat_type, 2L)
+    }
+    
+    if (integration_method == "harmony") {
+        a <- get_args_list(...)
+        a$integration_method <- NULL
+        # this function needs error handling or it locks the console
+        res <- tryCatch({
+            do.call(.lab_transfer_harmony, a)
+        }, error = function(e) {
+            stop(wrap_txtf("labelTransfer: harmony:\n%s", e$message), 
+                 call. = FALSE)
+        })
+        return(res)
+    }
+    
+    # NSE vars
+    temp_name <- cell_ID <- temp_name_prob <- NULL
+    
+    su1 <- set_default_spat_unit(x, spat_unit = spat_unit[[1]])
+    su2 <- set_default_spat_unit(y, spat_unit = spat_unit[[2]])
+    ft1 <- set_default_feat_type(
+        x, spat_unit = su1, feat_type = feat_type[[1]]
+    )
+    ft2 <- set_default_feat_type(
+        y, spat_unit = su2, feat_type = feat_type[[2]]
     )
 
     # get data
     cx_src <- getCellMetadata(y,
-        spat_unit = spat_unit,
-        feat_type = feat_type,
+        spat_unit = su2,
+        feat_type = ft2,
         output = "data.table"
     )
     cx_tgt <- getCellMetadata(x,
-        spat_unit = spat_unit,
-        feat_type = feat_type,
+        spat_unit = su1,
+        feat_type = ft1,
         output = "data.table"
     )
     dim_coord <- getDimReduction(x,
-        spat_unit = spat_unit,
-        feat_type = feat_type,
+        spat_unit = su1,
+        feat_type = ft1,
         reduction = reduction,
         reduction_method = reduction_method,
         name = reduction_name,
@@ -3479,8 +3787,8 @@ setMethod("labelTransfer", signature(x = "giotto", y = "giotto"), function(
 
     if (return_gobject) {
         x <- addCellMetadata(x,
-            spat_unit = spat_unit,
-            feat_type = feat_type,
+            spat_unit = su1,
+            feat_type = ft1,
             new_metadata = cx_tgt,
             by_column = TRUE,
             column_cell_ID = "cell_ID"
@@ -3490,6 +3798,9 @@ setMethod("labelTransfer", signature(x = "giotto", y = "giotto"), function(
         return(cx_tgt)
     }
 })
+
+
+# ** giotto, missing ####
 
 #' @rdname labelTransfer
 #' @export